Move oresat mission arg handling into argparse
This commit enhances the oresat mission argument in scripts with
awareness of valid missions and the currently active mission. It now:
 - Lists valid values in the help message
 - Generates the list of valid values automatically
 - Picks an appropriate default argument (the currently active mission)
   automatically
 - Rejects invalid values in the argument parser itself

It tries to be similar but not compatible to the previous handling:
 - Accepts values of the form "oresat0.5" but also accepts values without
   the prefix, "0.5".
 - Is a flag instead of a positional argument. "--oresat 0.5" looks
   similar to the previous positional argument "oresat0.5". Since it
   picks an appropriate default value now, most uses can be elided, so
   the slightly longer option I felt was acceptable.

So far every script uses this argument, which would make it a candidate
for putting it in __main__ (the top level argument handler), but since
each script currently needs to be able to be run as a standalone as
well, I didn't see a good way of enabling that without duplicating the
option in each script as well. If it's in each script anyway then
there's not much point in having it at the top level.

To support this arg the Consts enum is introduced. It bundles short
names ("0.5"), nice names ("OreSat0.5", via __str__()) and IDs (2)
together as well as a handful of utilities for converting between
types. This grew out of OreSatId but was renamed because it no longer
is just an ID. It's a dataclass enum instead of an IntEnum, because I
don't believe the special properties of an IntEnum were being used.
IntEnum lets integer operations (arithmetic, ordered comparisons, etc.)
be used directly on the enum members but they wouldn't really make sense
for a mission ID. Maybe ordered comparisons to find out which mission
was first, but that isn't how it's been used so far. Consts can also be
extended with other info in the future, this current implementation is
the minimum set for the current needs.

Consts subsumes ORESAT_NICE_NAMES and OreSatId, and while
ORESAT_NICE_NAMES has been dropped, OreSatId has been left in for now as
it's part of the current public interface to oresat_configs. When a new
version of oresat_configs is released, users can convert to Consts, and
then OreSatId can be removed. It's not been marked as @deprecated to
reduce noise, I'll just fix things up when the release happens.

OreSatConfig() has also been enhanced to accepts multiple types, which
it then uses the Consts type conversion utils to coax the argument into
a Const.