Add card/beacon configs path to Consts
As an example of future extensions from the previous commit, this
extends Consts with the card and beacon config paths, meaning that
OreSatConfig doesn't need to have special dicts for each one.

It also adds some type information and shortens the names a bit. Since
for example ORESAT0_BEACON_CONFIG_PATH is in the oresat0 module, the
module prefix in the name is redundant, it can just be
BEACON_CONFIG_PATH. If it needs to be distinguished from other beacon
paths, the module name can be used as well: oresat0.BEACON_CONFIG_PATH.