Shorten subcommand names
When invoked as part of __main__, the script names don't need to be
long. Too long hinders readability and is extra effort to write.
Details about the subcommand operation can be found in the adjacent
description.

This also drops the *_PROG constants, since at this point they only had
one user and don't need to be accessed outside the script module.