Add list-cards script subcommand
This subcommand lists the known cards for a given oresat mission. Many
other subcommands take a card as an argument and knowing in advance
which strings work as valid cards is not obvious. All the info is in
cards.csv of course, and this just prints from that, but cards.csv may
not be easily accessible. I had considered putting the valid cards in
the help text to cards arguments, but there are too many to be easily
readable.

I chose not to make this also a standalone script because it's only
informational. I could go either way on it though, not strongly
attached.

This uses the tabulate package for formatting tables, which has been
added to the requirements, along with its type information.

Since this script only uses card.csv and doesn't need the OD or other
heavyweight info, the csv to Cards feature has been pulled out into its
own function. This makes startup of list-cards much quicker.