Ensure yaml matches dataclass types
This adds a series of tests for the conversion of yaml config files to
their corresponding dataclass types, along with fixes to discovered
errors.

The main error is really one of yaml's fault. In yaml 1.1 bools are
not only true/false, but also yes/no, y/n, and, specific to here,
on/off. See https://yaml.org/type/bool.html for more. pyyaml will
convert these types to python's bool: True or False. Where on and off
are used in the configs, they are intended to be strings so quoting
needs to be applied to keep the string value.

As an added bonus dataclasses_json's from_dict() method is type aware,
so when it gets a mixed key types dict, it will implicitly convert the
now bool values to string "True" or "False", as that's what the
CardConfig specifies for the relevant field. For example if you load
c3.yaml into a CardConfig c then:

 c.objects[-2].generate_subindexes.value_descriptions ==
{'False': 0, 'boot': 1, 'True': 2, 'error': 3, 'not_found': 4, 'dead': 255}

Where we would expect {'off': 0, ..., 'on': 2, ...}.

The tests introduce the dacite package, which does a similar from_dict
that dataclasses_json does, but is type aware and won't do implicit
conversion, raising an exception instead. It even has a strict checking
mode to ensure all entries are consumed, but using it will have to wait
for a later time.

An alternate solution I considered was switching to yaml 1.2, which
defines bools as strictly true or false, but it turns out pyyaml doesn't
yet support 1.2 despite it being out since 2009. In fact, it doesn't
look like the pyyaml project is very alive any more. The last update was
in 2020 and its ticket tracker is slowly fading out.