Memoize CardConfig.from_yaml()
A recent profile shows that CardConfig.from_yaml() takes up about 80% of
the total time spent in OreSatConfig(). Given that it takes ~50 seconds
to construct on the C3, that's ~40 seconds creating CardConfigs.

Profiling shows that CardConfig.from_yaml() is called 35 times, but
since there's only 16 yaml files and each CardConfig is supposed to
mirror each yaml file, we only need to create at most 16 unique
CardConfigs. Most of the duplicates come from _load_configs() loading
the fw_ and sw_ common.yaml files repeatedly. We could try to carefully
construct only exactly the CardConfigs needed, but there's an easy trick
to get most of the way there: memoization (not to be confused with
memorization).

Memoization is the technique of caching the output of expensive
functions and then looking up the result later when the function is
called again instead of the expensive calculation. Python makes
memoization very easy with the functools.cache decorator.

Using @cache on CardConfig.from_yaml() shows that it now only gets
called 11 times, bringing OreSatConfig() construction time down to ~9
seconds.