Add descriptive cards.csv column check
To ensure they line up with the definition of the Card dataclass. This
makes it abundantly clear what the problem is when cards.csv and Card
gets out of sync.

It was a tossup between ValueError and TypeError. I really didn't want
to do a custom error just for this, and ValueError is the default
choice, but since it's about either the type of Card being wrong or the
"type" of the csv being wrong, I felt like I could get away with
TypeError here.