Ensure pyyaml has libyaml C bindings
For performance we want to only use the C bindings version of pyyaml
(and also the docs mention that there are some unspecified minor
differences in parsing). On x86, with the requisite library installed,
this works by default, but on ARM systems (like the C3 Octavo) the
wheels provided by PyPI aren't built with the bindings. This means we
have to build them from source, which thankfully pip makes not onerously
difficult, but some amount of deviation from the standard install is
required.

This makes `OreSatConfigs()` about 2/3 faster, in my testing I saw a
reduction from 9s to 3s.

This also updates the Github actions to stop deprecation warnings and
uses the cache feature for pip to make it go a bit faster.