Refresh docs for __main__.py
This makes it look better in the pydoc output. This is why _SCRIPTS was
made explicitly internal and the the oresat_configs() function was
removed, they don't need to show up in the module docs.

I will say it's pretty weird to have meaningful module docs for
__main__. In a survey of installed packages on my system none of them
have meaningful module docstrings.