Update pyproject.toml with Python 3.9, dacite, direnv
This is a handful of minor fixups to pyproject.toml. Since we we have
only been testing with Python 3.9 the minimum supported version has been
de facto that due to later language features sneaking in. If there are
people still on 3.7 I can work with them to set up pyenv to manage a
local install of python 3.9.

This also completes the purge of dataclasses-json

Finally it tells pylama to ignore my personal virtualenv tooling, I
didn't immediately see an easy way to do that only locally.