Fix installed scripts, add test
In 3404899 as part of a cleanup I removed the oresat_configs() function
in __main__.py. It turns out that this function was being used, as the
entry point for the installed oresat-configs script.

In order to not break it again I've added a CI test that runs the
expected installed scripts. It's not an exhaustive system test, more of
a simple smoke test to make sure they don't crash.