Fix overlays overwriting later loaded missions
In 037232 the output of CardConfig.from_yaml() was memoized. That is, as
long as the inputs (the yaml path) is the same it will return the
already constructed CardConfig instead of building a new one. This is
fine if an application only loads a single mission but if all missions
are loaded starting with OreSat0, OreSat0 will apply an overlay to the
returned CardConfig, modifying it in place. This modification then,
because of caching, persists when later missions try to also load that
same CardConfig.

Specifically, in some cases the OreSat0 battery overlay could be applied
to OreSat0.5/OreSat1. oresat-c3-software only loaded OreSat0.5 and so it
got the correct version of the battery, but gen_beacon_rst.py in
docs/scripts, because it generates docs for all three missions, was
giving OreSat0's int16 battery temperatures for OreSat0.5/OreSat1
instead of the correct int8.