Add yamllint
yamllint is a linter for yaml (who would have thought) which will help
with catching indentation, excess spaces, and things like that. This
specific pass caught
- A couple of mis-indented lists
- Two spaces before inline comments, like how black does
- Some lines greater than 100 characters
- Excess spaces around some syntax

I disabled warning on missing document start because they're optional
and we're fairly unambiguous about what we're doing. I also changed two
floats to scientific notation, but pyyamls parser for that is a bit
annoying. It will interpret them as a float and not a string only under
the conditions where it has a decimal point and a sign on the exponent.