Determine yaml configs from card info
cards.csv has all the info to determine what yaml configs to load for
each card. This enhances class Card to find that information and then
uses the new locations throughout. This allows for the removal of
filename constants and ConfigPaths dicts in the various __init__s,
removing one more location that needs to be updated when adding a new
card.

There were a few special cases in the file names that could be handled
by renaming them or creating an alias table, see Card.__post_init__ for
those cases.