OreSatConfig can be called with None/no argument
Most invocations of OreSatConfigs at this point are with
Mission.default() as the argument so we might as well be nice and do it
here instead of asking each user.

mypy got a little picky about types so this change ended up being bigger
than I expected but most of it is just mission -> self.mission.