Refactor attr_lines
`attr_lines` was generating initializers for each of the three OD types
in slightly different ways. This pulls that block out into its own
function so that fixes always happen to all three types, simplifies the
generating a bit, and fixes a few edge cases:
- Removed whitespace at the end of VISIBLE_STRING
- Use '\0' instead of 0 for null in VISIBLE_STRING
- VISIBLE_STRING would generate strings ("") instead of chars ('') as
  the array members for ODVariable and ODArray
- _SKIP_INDEXES are always skipped, previously ODRecords were not.

Checking over our current generated configs the only difference in
generated output is using '\0' instead of 0 for VISIBLE_STRINGs and
removing the single space at the end of those lines.